+++ Step Tracker/Charts/Chart Custom Views/ChartAnnotationView.swift
