+++ Step Tracker/Charts/Chart Utilities/ChartDataTypes.swift
