--- Step Tracker/Charts/ChartHelper.swift
@@ -1,21 +1,0 @@
-//
-//  ChartHelper.swift
-//  Step Tracker
-//
-//  Created by Jeremy Cinq-Mars on 2024-07-09.
-//
-
-import Foundation
-
-struct ChartHelper {
-    static func convert(data: [HealthMetric]) -> [DateValueChartData] {
-        data.map { .init(date: $0.date, value: $0.value)}
-    }
-    
-    static func parseSelectedData(from data: [DateValueChartData], in selectedDate: Date?) -> DateValueChartData? {
-        guard let selectedDate else { return nil }
-        return data.first {
-            Calendar.current.isDate(selectedDate, inSameDayAs: $0.date)
-        }
-    }
-}
