+++ Step Tracker/Charts/Charts/StepBarChart.swift
