+++ Step Tracker/Charts/Charts/StepPieChart.swift
@@ -91,5 +91,5 @@ struct StepPieChart: View {
 }
 
 #Preview {
-    StepPieChart(chartData: ChartMath.averageWeekdayCount(for: MockData.steps))
+    StepPieChart(chartData: ChartHelper.averageWeekdayCount(for: MockData.steps))
 }
